Visualization of tracing in query result
Why:
Tracing [1] is optionally returned from GraphQL servers but is hard to interpret without visualization.

What:
Tracing visualization rendered in the footer of the GraphiQL result window. Each tracing.execution.resolvers entry is rendered as a row with a bar together with resolver path label and duration.

Testing:
- Chrome Browser, Version 71.0.3578.98 (Official Build) (64-bit)
- Mac OSX 10.12.6
- GraphQL-java server with and without tracing turned on.

Left to do:
- Resizable panel where tracing visualization resides (currently max-height set to 300px to not consume all of the result window).
- Clear the tracing visualization when new query has been triggered.

[1] https://github.com/apollographql/apollo-tracing