Fully reload devtools when a page reload happens (#205)
* Fully reload devtools when a page reload happens

Right now when an application page reload is triggered, devtools
becomes out of synch with the newly refreshed application. This
means watched queries are no longer updated in the devtools panel,
until the devtools are fully closed / re-opened. To keep the
devtools in synch with the loaded application's Apollo Client
instance, this commit reloads the devtools application if a page
reload is triggered.

Fixes #108.

* Changelog update