Adjust "Local State" detection approach
The check to see which version of "Local State" handling the
installed version of Apollo Client supports, is broken. It is
leading devtools to believe that `apollo-link-state` is the only
supported state handling approach. It looks like this has been
broken for quite some time unfortunately.

This commit adjusts the approach for detecting newer local state
handling capabilities, and tweaks the code a bit to address
errors that occurred after now being able to properly run against
applications with newer AC 2.x versions installed.