Temporarily adjust the display of cache entities for AC3
We're working on a complete re-write of the AC devtools, so this
commit should be considered temporary.

Apollo Client's internal cache structure has changed between
AC2 and AC3. This means the `Inspector` code bits of devtools
(that drive the Cache panel) are no longer compatible with the
newer cache structure. The current `Inspector` relies on AC2's
internal cache structure to render a hierarchical view of the
cache, with links between different entities. Re-creating this
for AC3 will take time away from our newer devtools rewrite
work, so this commit focuses on making the minimal changes
needed to get the Cache panel working with AC3.

With this commit, AC3 cache contents are displayed in the
Cache panel, but contents are not properly linked (e.g. cache
references are not followed and linked together). Nested
entities are displayed as simple JSON dumps.