Make sure graphiql-explorer started queries are valid
When a new query is started via the graphiql-explorer (3rd party
plugin), it is currently being passed to graphiql to be run, with
an empty root selection set. This ends up breaking some of
graphiql-explorer's functionality, such as being able to rename a
query. This commit ensures that initial selection sets coming
from graphiql-explorer always include at least a `__typename`
field, to prevent the first query from breaking.