Remove __ref replacement when showing cache trees
While it useful to automatically replace __ref's with their
associated objects when showing various branches in the
cache tree, this hides the underlying structure of the cache
and could lead to a misunderstanding of how the cache works.
We might bring this functionality back in a future release, or
make it an option, but we're removing it for now.