Remove extra introspection query check
The extra check and swap to return a resolved empty object
via a Promise was causing valid results to be randomly
replaced with an empty object. This is especially noticeable
when running devtools against a production application. We'll
remove the empty object replacement code and live with the
extra introspection requests for now (since they are relatively
harmless).