Make sure fragment imports are resolved
This was a regression caused by trying to extract a string based
representation of an operation from a document AST's `source`
property, to avoid making an extra call to `print`. While this
works in a lot of cases, it does not handle source imports
properly as the `#import ...` statement is present in `source`.
This commit avoids using `source`, making sure we convert the
AST into a string using `print`, to handle imports properly.

Fixes #462