Modernize many of the packages and upgrade the code theme to something more Apollo-y (#994)
Upgrades to the latest versions of many packages including React and Apollo. As part of this upgrade, some packages were wholly replaced as they were either unmaintained/lightly maintained and/or were incompatible with React 18 (according to their peer dependencies). The code theme from the Apollo docs was also imported for GraphQL syntax highlighting to provide a more cohesive feel.