fix: check if handler exist before sending tab command (#1713)
* fix: check if handler exist before sending tab command

The devtool monkey patches the stop method, but the stop method can be
called in react strict mode twice. On the second call, the Map does
not have a handler and leads to reading id of undefined error.

This change enables back the type safety and make sure id exists
before calling clientTerminated